refactor: extract reports panel from PanelViews.tsx
Batch 6 of the per-panel split for issue #144: reports - the second
biggest remaining panel, isolated because of its internal role-branching
(not just its size). Only student_profile is left after this.

- New panels/ReportsPanel.tsx ({ currentUser, schools, students,
  reportsList } props). Internally branches on currentUser.role:
  State Admin (UserRole.ADMIN) gets a district drill-down sub-view,
  every other role gets the flat evaluation-list view. Both branches
  moved verbatim, byte-for-byte.
- Moves in expandedDistRpt useState (previously top-level, used only
  by the State Admin sub-view).
- Uses the already-exported handleDownloadPDF from pdfReportGenerator.ts
  (student_profile still uses it too from PanelViews.tsx directly,
  unaffected).
- PanelViews.tsx's 1 if-block replaced with a one-line router call.
- Unused-import cleanup: removed MapPin, School as SchoolIcon (both
  confirmed dead - reports was their only remaining usage after prior
  batches), and the DISTRICT_NAMES import (same reason).

Verified: tsc --noEmit clean, vite build succeeds, byte-diff of the
moved block, plus a structural re-read of the whole extracted file
before running tsc (given the off-by-one caught in the previous PR).
Live click-through covering both role branches against a scratch
backend with 2 real seeded evaluation reports: Punjab State Coordinator
(ADMIN role) - drill-down into Ludhiana district works, expandedDistRpt
toggles correctly, per-school report counts match. Superadmin - flat
evaluation list renders both reports (one resolves to a real student
name, the other legitimately shows "Unknown" because /api/students is
capped at 1000 for Superadmin by pre-existing design and that student
falls outside the returned subset - confirmed unrelated to this PR by
checking the same cap already showed up in an earlier PR's Analytics
panel test), and the Download PDF Report button opens a real print-preview
tab titled "Assessment Report - Sara Singh", confirming handleDownloadPDF
threads through correctly.
